requirements.txt
PyGithub>=2.1.1 
openai>=1.0.0 
requests>=2.31.0 
python-dotenv>=1.0.0 
dspy-ai>=2.0.0 
pyyaml>=6.0.1 
